Move musl reference tests into its own file
