+++ .gitignore
@@ -4,5 +4,4 @@
 /math/src
 /math/target
 /target
-/tests
 Cargo.lock
